requirements.txt
pytorch_ignite==0.1.1
pytest==3.10.0
blackhc.laaos~=0.1
blackhc.notebook
blackhc.progress_bar
numpy>=1.15.4
typing==3.6.6
dataclasses==0.6
pytest-benchmark
sklearn
numba
matplotlib
